+++ .vscode/settings.json
@@ -1,4 +1,4 @@
 {
-  "python.pythonPath": "env/bin/python3.6",
+  "python.pythonPath": "/usr/bin/python3",
   "python.formatting.provider": "black"
 }
